+++ CHANGELOG.md
@@ -11,12 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 * Changed xc3_gltf to support multiple input files and multiple animation files.
+* Lowered resolution of low resolution textures to better match original file sizes for wimdo texture export.
 
 ### Fixed
 * Fixed a potential crash in xc3_viewer when drawing empty skeletons with the `--bones` option.
 * Fixed an issue where wimdo models would not load the `.chr` or `.arc` file with the full base skeleton.
 * Fixed an issue where glTF export would generate an invalid emissive texture reference in some cases.
 * Fixed an issue where JPEG export would not work with xc3_tex.
+* Fixed an issue where textures would load invalid mipmap data in some cases.
 
 ## 0.16.0 - 2025-01-03
 ### Added